HomeStay-MCP/pyproject.toml
[project]
name = "homestay-mcp"
version = "0.1.0"
description = "MCP Server for Homestay Management and Filtering"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "mcp[cli]>=1.12.3",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "uvicorn>=0.23.2",
    "motor>=3.3.0",
    "pymongo>=4.6.0",
]
